Use Float for the warmup/runtime.
Warmup/run times such as 2.5 seconds are perfectly valid, as such we have to
use a Float for the column's type.

+++ db/migrate/20140618130126_create_schema.rb
@@ -8,9 +8,11 @@ Sequel.migration do
       String :version
       String :language
 
-      Integer :warmup
-      Integer :runtime
+      Float :warmup
+      Float :runtime
+
       Integer :iterations
+
       Float :iteration_time
       Float :iterations_per_second
